Spring Security code
updated and added Spring Security code

+++ src/main/java/com/coderscampus/flightTrack/domain/Role.java
@@ -16,7 +16,7 @@ public class Role implements GrantedAuthority{
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name="role_id")
-	private Integer roleId;
+	private Long roleId;
 	
 	private String authority;
 	
@@ -26,7 +26,7 @@ public class Role implements GrantedAuthority{
 	public Role(String authority) {
 		this.authority = authority;
 	}
-	public Role(Integer roleId, String authority) {
+	public Role(Long roleId, String authority) {
 		this.roleId = roleId;
 		this.authority = authority;
 	}
@@ -37,10 +37,10 @@ public class Role implements GrantedAuthority{
 	public void setAuthority(String authority) {
 		this.authority = authority;
 	}
-	public Integer getRoleId() {
+	public Long getRoleId() {
 		return this.roleId;
 	}
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
     	this.roleId = roleId;
     }
 }
